requirements.txt
discord.py
PyNaCl==1.4.0
async-timeout==3.0.1
rule34~=1.7.4
dnspython==2.0.0

motor~=2.4.0
asyncio~=3.4.3
python-dateutil~=2.8.1
DiscordUtils~=1.2.9.1
aiohttp~=3.7.4
emojis~=0.6.0